moves types to server folder
