+++ src/routes/album/[albumId]/ArtistReviews/ArtistReviews.tsx
@@ -2,7 +2,7 @@ import { component$ } from "@builder.io/qwik";
 import type { Session } from "next-auth";
 import { ReviewList } from "~/modules/ReviewList/ReviewList";
 import type { findAlbum } from "~/server/album";
-import type { AsyncReturnValue } from "~/utils/types";
+import type { AsyncReturnValue } from "~/server/types";
 import { deleteReviewAction } from "../layout";
 
 type Props = {
